+++ docs/examples/code/fill_and_submit_web_form_crawler.py
@@ -2,7 +2,8 @@ import asyncio
 from urllib.parse import urlencode
 
 from crawlee import Request
-from crawlee.http_crawler import HttpCrawler, HttpCrawlingContext
+from crawlee.http_crawler import HttpCrawler
+from crawlee.static_content_crawler._static_crawling_context import HttpCrawlingContext
 
 
 async def main() -> None:
