+++ src/crawlee/beautifulsoup_crawler/__init__.py
@@ -1,5 +1,5 @@
 try:
-    from ._beautifulsoup_crawler import BeautifulSoupCrawler, BeautifulSoupParser
+    from ._beautifulsoup_crawler import BeautifulSoupCrawler
     from ._beautifulsoup_crawling_context import BeautifulSoupCrawlingContext
 except ImportError as exc:
     raise ImportError(
@@ -7,4 +7,4 @@ except ImportError as exc:
         "For example, if you use pip, run `pip install 'crawlee[beautifulsoup]'`.",
     ) from exc
 
-__all__ = ['BeautifulSoupCrawler', 'BeautifulSoupCrawlingContext', 'BeautifulSoupParser']
+__all__ = ['BeautifulSoupCrawler', 'BeautifulSoupCrawlingContext']
