+++ src/crawlee/parsel_crawler/_parsel_parser.py
@@ -6,7 +6,7 @@ from typing_extensions import override
 
 from crawlee._utils.blocked import RETRY_CSS_SELECTORS
 from crawlee.http_clients import HttpResponse
-from crawlee.http_crawler import BlockedInfo, StaticContentParser
+from crawlee.static_content_crawler._static_content_parser import BlockedInfo, StaticContentParser
 
 
 class ParselParser(StaticContentParser[Selector]):
