+++ src/crawlee/static_content_crawler/__init__.py
@@ -1,0 +1,11 @@
+from ._static_content_crawler import StaticContentCrawler
+from ._static_content_parser import BlockedInfo, StaticContentParser
+from ._static_crawling_context import HttpCrawlingContext, ParsedHttpCrawlingContext
+
+__all__ = [
+    'BlockedInfo',
+    'ParsedHttpCrawlingContext',
+    'HttpCrawlingContext',
+    'StaticContentCrawler',
+    'StaticContentParser',
+]
