+++ tests/unit/http_crawler/test_http_crawler.py
@@ -20,8 +20,7 @@ if TYPE_CHECKING:
     from collections.abc import AsyncGenerator, Awaitable
 
     from crawlee.http_clients._base import BaseHttpClient
-    from crawlee.http_crawler._http_crawling_context import HttpCrawlingContext
-
+    from crawlee.static_content_crawler._static_crawling_context import HttpCrawlingContext
 
 # Payload, e.g. data for a form submission.
 PAYLOAD = {
